+++ src/Ext/SWType/Body.cpp
@@ -48,7 +48,7 @@ void SWTypeExt::ExtData::Serialize(T& Stm)
 		.Process(this->SW_FirstClickRestartsTimer)
 		.Process(this->SW_FirstClickRestartsTimer_Cost)
 		.Process(this->SW_FirstClickRestartsTimer_AutoFire)
-		.Process(this->SW_FirstClickRestartsTimer_AutoFire_ForceDetonations)
+		.Process(this->SW_FirstClickRestartsTimer_ForceDetonations)
 		.Process(this->Message_InsufficientFunds)
 		.Process(this->Message_RestartedTimer)
 		.Process(this->EVA_RestartedTimer)
@@ -193,7 +193,7 @@ void SWTypeExt::ExtData::LoadFromINIFile(CCINIClass* const pINI)
 	this->Message_RestartedTimer.Read(exINI, pSection, "Message.RestartedTimer");
 	this->EVA_RestartedTimer.Read(exINI, pSection, "EVA.RestartedTimer");
 	this->SW_FirstClickRestartsTimer_AutoFire.Read(exINI, pSection, "SW.FirstClickRestartsTimer.AutoFire");
-	this->SW_FirstClickRestartsTimer_AutoFire_ForceDetonations.Read(exINI, pSection, "SW.FirstClickRestartsTimer.AutoFire.ForceDetonations");
+	this->SW_FirstClickRestartsTimer_ForceDetonations.Read(exINI, pSection, "SW.FirstClickRestartsTimer.ForceDetonations");
 }
 
 void SWTypeExt::ExtData::LoadFromStream(PhobosStreamReader& Stm)